requirements.txt
bcrypt==3.1.3
cffi==1.10.0
click==6.6
flake8==2.5.4
flake8-docstrings==0.2.6
Flask==0.12.2
Flask-Cache==0.13.1
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==0.23
marshmallow==2.7.3
mccabe==0.4.0
mimerender==0.6.0
pep257==0.7.0
pep8==1.7.0
pycparser==2.14
pyflakes==1.0.0
python-dateutil==2.5.3
python-mimeparse==1.5.2
pytz==2016.4
six==1.10.0
SQLAlchemy==1.1.13
Unidecode==0.4.19
Werkzeug==0.11.10
Flask-HTTPAuth==3.2.2
selenium==2.53.6
Flask-Cors==3.0.3
PyYAML==3.12
